+++ include/COMMON/fileContainer.h
@@ -5,19 +5,19 @@
 
 typedef struct _fileContainer
 {
-    uint64_t size;
+    size_t size;
     void *content;
     char *name;
 } FileContainer;
 
-int containerInit(uint64_t size, void *content, const char *name, FileContainer *fc);
+int containerInit(size_t size, void *content, const char *name, FileContainer *fc);
 void destroyContainer(FileContainer *fc);
-int serializeContainer(FileContainer fc, void **buf, uint64_t *size);
-FileContainer deserializeContainer(void *buf, uint64_t size);
+int serializeContainer(FileContainer fc, void **buf, size_t *size);
+FileContainer deserializeContainer(void *buf, size_t size);
 
-int serializeContainerArray(FileContainer *fc, uint64_t n, uint64_t *size, void **buf);
-FileContainer *deserializeContainerArray(void *buf, uint64_t size, uint64_t *n);
+int serializeContainerArray(FileContainer *fc, size_t n, size_t *size, void **buf);
+FileContainer *deserializeContainerArray(void *buf, size_t size, size_t *n);
 
-uint64_t calcSize(FileContainer fc);
+size_t calcSize(FileContainer fc);
 
 #endif
