+++ include/SERVER/connstate.h
@@ -21,9 +21,9 @@ void connStateDestroy(ConnState *state);
 int conn_openFile(char *path, int flags, FileContainer **fcs, int *fcsSize, ConnState state);
 int conn_closeFile(const char *path, ConnState state);
 
-int conn_readFile(const char *path, void **buf, uint64_t size, ConnState state);
-int conn_writeFile(const char *path, void *buf, uint64_t size, FileContainer **fcs, int *fcsSize, ConnState state);
-int conn_appendFile(const char *path, void *buf, uint64_t size, FileContainer **fcs, int *fcsSize, ConnState state);
+int conn_readFile(const char *path, void **buf, size_t size, ConnState state);
+int conn_writeFile(const char *path, void *buf, size_t size, FileContainer **fcs, int *fcsSize, ConnState state);
+int conn_appendFile(const char *path, void *buf, size_t size, FileContainer **fcs, int *fcsSize, ConnState state);
 
 int conn_readNFiles(int N, FileContainer **fcs, ConnState state);
